+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/bean/dto/FundChangeRecordInfo.java
@@ -4,6 +4,7 @@ import com.dwarfeng.subgrade.stack.bean.dto.Dto;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 
 import java.math.BigDecimal;
+import java.util.Date;
 
 /**
  * 资金变更记录信息。
@@ -13,21 +14,37 @@ import java.math.BigDecimal;
  */
 public class FundChangeRecordInfo implements Dto {
 
-    private static final long serialVersionUID = -5940553395011650098L;
+    private static final long serialVersionUID = -5892794110185375481L;
 
     private LongIdKey accountBookKey;
     private BigDecimal delta;
     private String changeType;
     private String remark;
 
+    /**
+     * 发生日期。
+     *
+     * <p>
+     * 发生日期指的是资金变更发生的日期，可以称为变更日期。
+     *
+     * <p>
+     * 该值可以为 <code>null</code>，表示发生日期为当前系统日期。
+     *
+     * @since 1.5.0
+     */
+    private Date happenedDate;
+
     public FundChangeRecordInfo() {
     }
 
-    public FundChangeRecordInfo(LongIdKey accountBookKey, BigDecimal delta, String changeType, String remark) {
+    public FundChangeRecordInfo(
+            LongIdKey accountBookKey, BigDecimal delta, String changeType, String remark, Date happenedDate
+    ) {
         this.accountBookKey = accountBookKey;
         this.delta = delta;
         this.changeType = changeType;
         this.remark = remark;
+        this.happenedDate = happenedDate;
     }
 
     public LongIdKey getAccountBookKey() {
@@ -62,6 +79,14 @@ public class FundChangeRecordInfo implements Dto {
         this.remark = remark;
     }
 
+    public Date getHappenedDate() {
+        return happenedDate;
+    }
+
+    public void setHappenedDate(Date happenedDate) {
+        this.happenedDate = happenedDate;
+    }
+
     @Override
     public String toString() {
         return "FundChangeRecordInfo{" +
@@ -69,6 +94,7 @@ public class FundChangeRecordInfo implements Dto {
                 ", delta=" + delta +
                 ", changeType='" + changeType + '\'' +
                 ", remark='" + remark + '\'' +
+                ", happenedDate=" + happenedDate +
                 '}';
     }
 }