+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/bean/entity/FundChange.java
@@ -14,21 +14,39 @@ import java.util.Date;
  */
 public class FundChange implements Entity<LongIdKey> {
 
-    private static final long serialVersionUID = -306278124499756731L;
+    private static final long serialVersionUID = -8814660888399626547L;
 
     private LongIdKey key;
     private LongIdKey accountBookKey;
     private BigDecimal delta;
     private String changeType;
+
+    /**
+     * 发生日期。
+     *
+     * <p>
+     * 发生日期指的是资金变更发生的日期，可以称为变更日期。
+     */
     private Date happenedDate;
     private String remark;
+
+    /**
+     * 记录日期。
+     *
+     * <p>
+     * 记录日期指的是资金变更被记录的日期。<br>
+     * 当资金变更被创建或被更新时，该值通常取当前的系统时间。
+     *
+     * @since 1.5.0
+     */
+    private Date recordedDate;
 
     public FundChange() {
     }
 
     public FundChange(
             LongIdKey key, LongIdKey accountBookKey, BigDecimal delta, String changeType, Date happenedDate,
-            String remark
+            String remark, Date recordedDate
     ) {
         this.key = key;
         this.accountBookKey = accountBookKey;
@@ -36,6 +54,7 @@ public class FundChange implements Entity<LongIdKey> {
         this.changeType = changeType;
         this.happenedDate = happenedDate;
         this.remark = remark;
+        this.recordedDate = recordedDate;
     }
 
     @Override
@@ -88,6 +107,14 @@ public class FundChange implements Entity<LongIdKey> {
         this.remark = remark;
     }
 
+    public Date getRecordedDate() {
+        return recordedDate;
+    }
+
+    public void setRecordedDate(Date recordedDate) {
+        this.recordedDate = recordedDate;
+    }
+
     @Override
     public String toString() {
         return "FundChange{" +
@@ -97,6 +124,7 @@ public class FundChange implements Entity<LongIdKey> {
                 ", changeType='" + changeType + '\'' +
                 ", happenedDate=" + happenedDate +
                 ", remark='" + remark + '\'' +
+                ", recordedDate=" + recordedDate +
                 '}';
     }
 }